REFACTOR(#005): ADD A COMPONENT TO REPRESENT THE FILES ASSOCIATED WITH A COURSE
A function to check access to materials has also been created.
